pyproject.toml
[tool.poetry]
name = "venture"
version = "1.1.0"
description = "Rofi / Wofi based project selector"
authors = ["Sean Collings <seanrcollings@gmail.com>"]
license = "MIT"
packages = [
    {include = "venture", from = "src"}
]


[tool.poetry.dependencies]
python = "^3.9"
arc-cli = "^2.2.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.800"
pylint = "^2.7.0"


[tool.poetry.scripts]
venture = "venture.__main__:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
